update delete in brand and category

+++ backend/src/app/modules/brand/brand.controller.ts
@@ -7,9 +7,25 @@ import { IBrand } from "./brand.interface";
 import pick from "../../../shared/pick";
 import { paginationFields } from "../../../constants/pagination";
 import { brandFilterableFields } from "./brand.constants";
+import { Brand } from "./brand.model";
+import { deleteFile, getFileUrl } from "../../../helpers/fileHandlers";
+import ApiError from "../../../errors/ApiError";
 
 const createBrand = catchAsync(async (req: Request, res: Response) => {
   const { ...brandData } = req.body;
+
+  const existingBrand = await Brand.findOne({ name: brandData.name });
+  if (existingBrand) {
+    if (req.file) {
+      deleteFile(req.file.filename);
+    }
+    throw new ApiError(httpStatus.BAD_REQUEST, "Brand already exists");
+  }
+
+  if (req.file) {
+    brandData.logo = getFileUrl(req.file.filename);
+  }
+
   const result = await BrandService.createBrand(brandData);
 
   sendResponse<IBrand>(res, {
@@ -49,8 +65,22 @@ const getSingleBrand = catchAsync(async (req: Request, res: Response) => {
 
 const updateBrand = catchAsync(async (req: Request, res: Response) => {
   const { id } = req.params;
-  const updatedData = req.body;
-  const result = await BrandService.updateBrand(id, updatedData);
+  const { ...updateData } = req.body;
+
+  const existingBrand = await Brand.findById(id);
+  if (!existingBrand) {
+    throw new ApiError(httpStatus.NOT_FOUND, "Brand not found");
+  }
+
+  if (req.file) {
+    if (existingBrand.logo) {
+      const oldFilename = existingBrand.logo.split("/").pop();
+      deleteFile(oldFilename ?? "");
+    }
+    updateData.logo = getFileUrl(req.file.filename);
+  }
+
+  const result = await BrandService.updateBrand(id, updateData);
 
   sendResponse<IBrand>(res, {
     statusCode: httpStatus.OK,
@@ -62,6 +92,17 @@ const updateBrand = catchAsync(async (req: Request, res: Response) => {
 
 const deleteBrand = catchAsync(async (req: Request, res: Response) => {
   const { id } = req.params;
+
+  const brand = await BrandService.getSingleBrand(id);
+  if (!brand) {
+    throw new ApiError(httpStatus.NOT_FOUND, "Brand not found");
+  }
+
+  if (brand.logo) {
+    const filename = brand.logo.split("/").pop();
+    deleteFile(filename ?? "");
+  }
+
   const result = await BrandService.deleteBrand(id);
 
   sendResponse<IBrand>(res, {
